Refactored dwarf.cc into many separate files.
As a monolith it was getting unmanageable
