+++ src/util.h
@@ -19,6 +19,7 @@
 
 #include "absl/numeric/int128.h"
 #include "absl/strings/string_view.h"
+#include "absl/strings/substitute.h"
 
 namespace bloaty {
